+++ R/add_labels.R
@@ -15,7 +15,7 @@ add_labels <- function(data) {
         labelled::set_label_attribute("Parkinson's disease"),
       `FXTAS Stage` = .data$`FXTAS Stage` |>
         labelled::set_label_attribute("FXTAS Stage"),
-      `Ataxia: severity*` = .data$`Ataxia: severity*` |>
+      `ataxia severity*` = .data$`ataxia severity*` |>
         labelled::set_label_attribute("Ataxia severity"),
       `MMSE total score*` = .data$`MMSE total score*` |>
         labelled::set_label_attribute("MMSE total score"),
@@ -34,6 +34,20 @@ add_labels <- function(data) {
         labelled::set_label_attribute(
           "RTI Five-choice movement time"
         ),
+      `Tandem Walk` =
+        .data$`Tandem Walk` |>
+        labelled::set_label_attribute("tandem walk"),
+      ataxia = .data$ataxia |> set_label_attribute("ataxia"),
+      `ataxia severity` =
+        .data$`ataxia severity` |>
+        set_label_attribute("ataxia severity"),
+      `head tremor` = .data$`head tremor` |> set_label_attribute("head tremor"),
+      `Corpus Callosum-Thickness` =
+        .data$`Corpus Callosum-Thickness` |>
+        set_label_attribute("Corpus Callosum Thickness"),
+      `any autoimmune disorder` =
+        .data$`any autoimmune disorder` |>
+        set_label_attribute("autoimmune diagnoses or symptoms")
     ) |>
     set_mri_var_labels()
 }