+++ R/clean_head_tremor_onset.R
@@ -1,10 +1,7 @@
-clean_head_tremor_onset <- function(dataset)
-{
+clean_head_tremor_onset <- function(dataset) {
   dataset |>
     dplyr::mutate(
-
-      `Head Tremor: Age of onset` =
-        .data$`Head Tremor: Age of onset` |> dplyr::recode("68-67" = "67.5"),
-      # "68-67" |> strsplit("-") |> sapply(FUN = function(x) median(as.numeric(x)))
+      `head tremor: age of onset` =
+        .data$`head tremor: age of onset` |> dplyr::recode("68-67" = "67.5"),
     )
 }