+++ R/format_demographics_table_as_flextable.R
@@ -1,16 +1,16 @@
 #' Format the demographics table as a flextable
 #'
 #' @param x a [gtsummary::gtsummary] object
-#' @param ... not used
+#' @inheritDotParams shared_flextable_settings
 #'
 #' @returns a [flextable::flextable]
-#' @dev
-format_demographics_table_as_flextable <- function(x, ...)
-{
+#' @export
+#' @keywords internal
+format_demographics_table_as_flextable <- function(x, ...) { # nolint: object_length_linter
   x |>
     gtsummary::as_flex_table() |>
     # update upper header label: '' to 'M vs. F'
-    shared_flextable_settings() |>
+    shared_flextable_settings(...) |>
     flextable::compose(
       part = "header",
       i = 1,
@@ -21,4 +21,3 @@ format_demographics_table_as_flextable <- function(x, ...)
     # reset p-value header column to bold font
     flextable::bold(part = "header", j = 6)
 }
-