+++ R/make_demographics_table.R
@@ -4,6 +4,8 @@
 #' `strata` and `vars`
 #' @param strata names of column variable, specified as [character]
 #' @param vars names of row variables, specified as [character]
+#' @param make_ft [logical] whether to convert to flextable
+#' @inheritDotParams format_demographics_table_as_flextable -x
 #' @inherit format_demographics_table_as_flextable return
 #' @export
 #'
@@ -22,7 +24,9 @@ make_demographics_table <- function(
       "FXTAS Stage",
       "CGG"
       # "ApoE")
-    )) {
+    ),
+    make_ft = TRUE,
+    ...) {
 
 
   data_to_use <- data |>
@@ -57,8 +61,13 @@ make_demographics_table <- function(
         "p-values represent tests for sex differences in",
         "distributions of characteristics, all CGG repeat levels."
       )
-    ) |>
-    format_demographics_table_as_flextable()
+    )
+
+  if (make_ft) {
+    to_return <-
+      to_return |>
+      format_demographics_table_as_flextable(...)
+  }
 
   return(to_return)