+++ tests/testthat/test-any_tremor.R
@@ -1,25 +1,25 @@
 test_that(
-  "`Any tremor (excluding head)` is consistent with the individual tremor variables",
+  "`any tremor (excluding head)` is consistent with the individual tremor variables",
   {
     skip_if_not(exists("trax_gp34_all"))
     library(dplyr)
     tremor_types = c(
-      "Intention tremor",
-      "Resting tremor",
-      "Postural tremor",
-      "Intermittent tremor"
+      "intention tremor",
+      "resting tremor",
+      "postural tremor",
+      "intermittent tremor hx"
     )
 
     inconsistent =
       trax_gp34_all |>
-      dplyr::filter(`Any tremor (excluding head)` %in%
+      dplyr::filter(`any tremor (excluding head)` %in%
                       "No tremors recorded",
                     if_all(all_of(tremor_types), is.na)) |>
       dplyr::select(any_of(c(
         "FXS ID",
         "Event Name",
         "Visit Date",
-        tremor_types, "Any tremor (excluding head)"
+        tremor_types, "any tremor (excluding head)"
       )))
 
     expect_equal(object = nrow(inconsistent), expected = 0)