+++ tests/testthat/test-make_demographics_table.R
@@ -3,18 +3,23 @@ test_that(
   code = {
     ft = test_data_v1 |>
       dplyr::filter(!is.na(CGG)) |>
-      make_demographics_table()
+      make_demographics_table(make_ft = FALSE)
 
-    html_file <- tempfile(fileext = ".html")
-    flextable::save_as_html(ft, path = html_file)
-    doconv::expect_snapshot_html(
-      vwidth = 1200, # based on what github actions seems to do
-      vheight = 744, # these get passed down to `webshot2::webshot()`
-      x = html_file,
-      name = "demographics_table",
-      engine = "testthat"
-    )
+    ft |> as.data.frame() |>
+      ssdtools:::expect_snapshot_data(name = "demographics-table")
   }
 )
 
-# need better tests here
+test_that(
+  desc = "results are consistent 2",
+  code = {
+    ft = test_data_v1 |>
+      dplyr::filter(!is.na(CGG)) |>
+      make_demographics_table(make_ft = TRUE, padding = 2)
+
+    ft |>
+      flextable::gen_grob(fit = "fixed") |>
+      vdiffr::expect_doppelganger(title = "demographics_table-2")
+
+  }
+)