+++ tests/testthat/test-plot_compact_pvd.R
@@ -5,7 +5,7 @@ test_that("`plot_compact_pvd()` produces consistent results", {
                              n_s = 3) |>
     plot_compact_pvd(scale_colors = c("red", "blue", "purple4"))
 
-  fig1 |> vdiffr::expect_doppelganger(title = "PVD")
+  fig1 |> vdiffr::expect_doppelganger(title = "pvd")
 
 
 })